Added header for class Projectile.
It's now commented out because I realized that class Projectile does not need the rotation functionalities of the class Entity it is inheriting from. Because of this, I need to consider moving said rotation functionalities to class Tank, which seems to make more sense according to the concepts of object oriented programming.