Fixed bullet shooting for main player's tank!
I'm not completely sure on the details of the problem; I plan to look into it more. What I do know is that the createBullets command was told to apply to any SceneNode of category SceneAirLayer, but both the Background and Air layers were of SceneAirLayer, so a bullet would be created twice for every one necessary time. However, I still do not know why the second, incorrect bullet was positioned awkwardly.