Changed bullet positioning technique.
Rather than use some complicated trigonometry to adjust the offset based on the shooting tank's rotation, a new SceneNode child attached to each tank is used to position bullets based on their offsets. Offset is no longer dependent on the size of the tank's bounding rectangle. Offset is given now in coordinates as opposed to multipliers.
The offsets were placed in the DataTables.