Bullets no longer kill each other--for now at least.
It turns out that allowing bullets (that is, your bullet and the enemy's bullet) to destory each other on contact makes it much harder to kill the hunting tanks. But, there is still a possibility that this feature may return in some way. Perhaps, bullets could have a stat that determines which one of them, if any, "wins" (doesn't get destroyed) upon collision with an opposing bullet.