Implemented SpawnBlocks(), but will have to remove the vertex array of BlockNode.
When I decided to implement all of the walls in one BlockNode, it didn't much occur to me how hard this would make checking collisions with those walls. It's normally easy to check collision with a vertex array of rectangles. The problem is that this method of checking the collision between one rectangle and any of the rectangles in the vertex array in a BlockNode breaks the SceneNode framework that uses getBoudingRect() to check collisions between SceneNodes, and there is no satisfactory way to check each rectangle in the vertex array using getBoundingRect(). So, each block will have to be made into its own block node, spawning as enemies do in SpawnBlocks().