Implemented an only somewhat satisfactory response by tanks for colliding with blocks.
The problem with the newly implemented collision response is that it unreasonably only permits (in most cases) a tank to move in one direction (away from the block) while colliding with a block. Something similar to what is done in World::adaptTanksPositions() to prevent boundary breaching would probably be better.