Made change to LevelData.txt for testing purposes.
A look at level one shows that I have been testing with multiple enemies in level one. In this test, I have been experiencing utter framerate failure (often dropping to 1 FPS). An optimization in collision detection must be made. The current detection method uses recursion, but this can be changed to iteration, a quadtree, or use of vectors to store each subclass of SceneNode or Entity (ex. Projectiles, Tanks, etc.).