Implemented counter for number of collision tests.
It seems that the number of collision tests is lower when using a quadtree than when using recursion, yet the framerte still sinks to below 1 FPS under both techniques. Something about the quadtree is taking too much time, and this could be having to reinsert the objects into the quadtree each frame (and consequently having to check whether to divide it for each insert of an object).