Reorganized the styles for the sections' background colors
The descriptions' table and the main menu's section tag each read different styles. This has fixed the issue of the table's white (or at least, a color close to white) background not working if the window is resized. However, in the main menu, the white background extends too far, so that must be fixed.